requirements.txt
Django==5.2.4
requests==2.32.3
python-telegram-bot==21.9
celery==5.4.0
redis==5.2.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
